pyproject.toml
[project]
name = "discord-1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "certifi>=2025.11.12",
    "discord-py>=2.6.4",
    "python-dotenv>=1.2.1",
]
